+++ pjf/P_fisica.java
@@ -5,7 +5,7 @@ public class P_fisica extends Cliente {
     String cpf;
 
     //construtor
-    public P_fisica(String cpf, String nome, String email, String endereço, int telefone, String dt_nascimento,
+    public P_fisica(String cpf, String nome, String email, String endereço, long telefone, String dt_nascimento,
             float renda_presumida, int RG, boolean inter_alugel, boolean inter_compra,double saldo, int id_imv_c) {
             super(nome, email, endereço, telefone, dt_nascimento, renda_presumida, RG,saldo,id_imv_c);
                 this.inter_alugel=inter_alugel;
